added new component for the currency icons in the navbar currency selector
